Extract shared audio playback into AudioSessionHandler
Voice recordings and audio files both drive the same player, so the handler no longer lives next to voice-only views.
